requirements.txt
opencv-python
Pillow
tensorflow
glob
itertools
